refactored abilities to better control load-order
