+++ config/env.rb
@@ -8,7 +8,9 @@ def require_files_that_match(pattern)
   Dir[File.dirname(__FILE__) + pattern].each {|file| require file }
 end
 Mongoid.load! File.dirname(__FILE__) + '/mongoid.yml'
+
 require_files_that_match('/../api/*.rb')
+require_files_that_match('/../abilities/*.rb')
 require_files_that_match('/../models/*.rb')
 
 require File.dirname(__FILE__) + '/../api.rb'
