+++ models/admin.rb
@@ -1,11 +1,12 @@
 require 'bcrypt'
-require 'openssl'
 
 module Basic
   module Models
     class Admin
       include Mongoid::Document
       include BCrypt
+
+      include Basic::Ability::Tokenizable
 
       attr_accessor :secret
 
@@ -38,18 +39,6 @@ module Basic
 
       def encrypt_password
         self.salt = Password.create(@secret)
-      end
-
-      def produce_token(ttl=(ENV['DEFAULT_SESSION_DURATION'] || 600))
-        digest = OpenSSL::Digest.new('sha256')
-        key = password
-        data = SecureRandom.uuid
-
-        token = OpenSSL::HMAC.hexdigest(digest, key, data)
-
-        puts "#{email}: HMAC(key=#{key} data=#{data} ttl=#{ttl}) => #{token}"
-
-        return self.tokens.create({value: token, expires_on: Time.now+ttl})
       end
     end
   end
